Restructure callout create buttons
